feat: 프로젝트 다이얼로그 분리
