+++ src/components/project-list/ProjectItem.tsx
@@ -1,21 +1,11 @@
 import { Circle, Edit3, LogOut, Trash } from "lucide-react"
 import { useState } from "react"
-import { useNavigate } from "react-router-dom"
-import { AlertDialog } from "@/components/common/AlertDialog"
-import { useToast } from "@/components/common/ToastContext"
+import { useProjectActions } from "@/hooks/project/useProjectActions"
 import { cn } from "@/lib/utils"
-import { projectApi } from "@/services/api/project-api"
-import { useProjectStore } from "@/stores/project-store"
 import type { Project } from "@/types/project"
 import { ProjectAvatars } from "./ProjectAvatars"
-import { ProjectDialog } from "./ProjectDialog"
+import { ProjectDialogs } from "./ProjectDialogs"
 import { ProjectToggle } from "./ProjectToggle"
-
-// 프로젝트 폼 데이터 타입
-type ProjectFormData = {
-  name: string
-  description: string
-}
 
 interface ProjectItemProps {
   project: Project
@@ -30,102 +20,15 @@ export default function ProjectItem({
   onToggle,
   onLeave,
 }: ProjectItemProps) {
-  const navigate = useNavigate() // 프로젝트 이동용
   const [isHovered, setIsHovered] = useState(false)
-  const { refreshProjects } = useProjectStore()
-  const { addToast } = useToast()
-
-  // 다이얼로그 상태들 - 한눈에 보기 쉽게
-  const [isToggleDialogOpen, setIsToggleDialogOpen] = useState(false)
-  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false)
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false)
-  const [isLeaveDialogOpen, setIsLeaveDialogOpen] = useState(false)
-  const [editLoading, setEditLoading] = useState(false)
-  const [deleteLoading, setDeleteLoading] = useState(false)
-
   const isHost = project.myRole === "OWNER"
 
-  // 핸들러 함수들 - 직관적으로 이해하기 쉽게
-  const handleProjectClick = () => {
-    navigate(`/projects/${project.id}/ide`)
-  } //라우터 추가
-
-  const handleToggleClick = () => setIsToggleDialogOpen(true)
-  const handleToggleConfirm = () => {
-    onToggle?.(project.id)
-    setIsToggleDialogOpen(false)
-  }
-
-  const handleEditClick = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setIsEditDialogOpen(true)
-  }
-  const handleEditConfirm = async (data: ProjectFormData) => {
-    setEditLoading(true)
-    try {
-      await projectApi.updateProject(project.id, {
-        projectName: data.name,
-        description: data.description,
-      })
-      setIsEditDialogOpen(false)
-      // 프로젝트 목록 새로고침
-      await refreshProjects()
-      // 성공 토스트 메시지
-      addToast({
-        type: "success",
-        title: "프로젝트가 성공적으로 수정되었습니다.",
-        duration: 3000,
-      })
-    } catch (error) {
-      console.error("프로젝트 수정 실패:", error)
-      // 실패 토스트 메시지
-      addToast({
-        type: "error",
-        title: "프로젝트 수정에 실패했습니다.",
-        duration: 3000,
-      })
-    } finally {
-      setEditLoading(false)
-    }
-  }
-
-  const handleDeleteClick = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setIsDeleteDialogOpen(true)
-  }
-  const handleDeleteConfirm = async () => {
-    setDeleteLoading(true)
-    try {
-      await projectApi.deleteProject(project.id)
-      setIsDeleteDialogOpen(false)
-      // 프로젝트 목록 새로고침
-      await refreshProjects()
-      // 성공 토스트 메시지
-      addToast({
-        type: "success",
-        title: "project deleted successfully.",
-        duration: 3000,
-      })
-    } catch (_error) {
-      // 실패 토스트 메시지
-      addToast({
-        type: "error",
-        title: "project deletion failed.",
-        duration: 3000,
-      })
-    } finally {
-      setDeleteLoading(false)
-    }
-  }
-
-  const handleLeaveClick = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setIsLeaveDialogOpen(true)
-  }
-  const handleLeaveConfirm = () => {
-    onLeave?.(project.id)
-    setIsLeaveDialogOpen(false)
-  }
+  // 커스텀 훅에서 모든 액션 관련 로직을 가져옴
+  const { dialogStates, loadingStates, handlers } = useProjectActions({
+    project,
+    onToggle,
+    onLeave,
+  })
 
   return (
     <>
@@ -134,7 +37,7 @@ export default function ProjectItem({
           "flex w-full cursor-pointer items-center justify-between border-[#e2e2e2] border-b bg-transparent px-3 py-2 text-left transition-colors",
           isHovered && "bg-[rgba(238,238,238,0.5)]"
         )}
-        onClick={handleProjectClick}
+        onClick={handlers.handleProjectClick}
         onMouseEnter={() => setIsHovered(true)}
         onMouseLeave={() => setIsHovered(false)}
         type="button"
@@ -151,7 +54,7 @@ export default function ProjectItem({
 
               {/* Toggle/Status Icon */}
               {isHost && onToggle ? (
-                <ProjectToggle checked={isToggled} onCheckedChange={handleToggleClick} />
+                <ProjectToggle checked={isToggled} onCheckedChange={handlers.handleToggleClick} />
               ) : (
                 <div className="flex h-3 w-3 items-center justify-center">
                   <Circle className="h-2 w-2 fill-green-500 text-green-500" />
@@ -171,7 +74,7 @@ export default function ProjectItem({
             {isHovered && isHost && (
               <button
                 className="flex h-3.5 w-3.5 items-center justify-center"
-                onClick={handleEditClick}
+                onClick={handlers.handleEditClick}
                 type="button"
               >
                 <Edit3 className="h-4 w-4 text-gray-400" />
@@ -186,7 +89,7 @@ export default function ProjectItem({
                 {isHost ? (
                   <button
                     className="flex h-3.5 w-3.5 items-center justify-center"
-                    onClick={handleDeleteClick}
+                    onClick={handlers.handleDeleteClick}
                     type="button"
                   >
                     <Trash className="h-4 w-4 text-gray-400" />
@@ -194,7 +97,7 @@ export default function ProjectItem({
                 ) : (
                   <button
                     className="flex h-3.5 w-3.5 items-center justify-center"
-                    onClick={handleLeaveClick}
+                    onClick={handlers.handleLeaveClick}
                     type="button"
                   >
                     <LogOut className="h-4 w-4 text-gray-400" />
@@ -207,59 +110,13 @@ export default function ProjectItem({
         </div>
       </button>
 
-      {/* 다이얼로그들 - 한 곳에서 모두 관리 */}
-
-      {/* 토글 상태 변경 다이얼로그 */}
-      <AlertDialog
-        cancelText="Cancel"
-        confirmText="Change"
-        description={`Current status: ${isToggled ? "ON" : "OFF"}`}
-        isOpen={isToggleDialogOpen}
-        onCancel={() => setIsToggleDialogOpen(false)}
-        onConfirm={handleToggleConfirm}
-        onOpenChange={setIsToggleDialogOpen}
-        showCloseButton={false}
-        title="Change status"
-        variant="default"
-      />
-
-      {/* 프로젝트 수정 다이얼로그 */}
-      <ProjectDialog
-        initial={{ name: project.name, description: project.description }}
-        isLoading={editLoading}
-        mode="edit"
-        onConfirm={handleEditConfirm}
-        onOpenChange={setIsEditDialogOpen}
-        open={isEditDialogOpen}
-      />
-
-      {/* 프로젝트 삭제 확인 다이얼로그 */}
-      <AlertDialog
-        cancelText="Cancel"
-        confirmText="Delete"
-        description="Are you sure you want to delete this project?"
-        isLoading={deleteLoading}
-        isOpen={isDeleteDialogOpen}
-        onCancel={() => setIsDeleteDialogOpen(false)}
-        onConfirm={handleDeleteConfirm}
-        onOpenChange={setIsDeleteDialogOpen}
-        showCloseButton={false}
-        title="Delete Project"
-        variant="destructive"
-      />
-
-      {/* 프로젝트 나가기 확인 다이얼로그 */}
-      <AlertDialog
-        cancelText="Cancel"
-        confirmText="Leave"
-        description="Are you sure you want to leave this project?"
-        isOpen={isLeaveDialogOpen}
-        onCancel={() => setIsLeaveDialogOpen(false)}
-        onConfirm={handleLeaveConfirm}
-        onOpenChange={setIsLeaveDialogOpen}
-        showCloseButton={false}
-        title="Leave Project"
-        variant="warning"
+      {/* 다이얼로그들 - 별도 컴포넌트로 분리 */}
+      <ProjectDialogs
+        dialogStates={dialogStates}
+        handlers={handlers}
+        isToggled={isToggled}
+        loadingStates={loadingStates}
+        project={project}
       />
     </>
   )
