+++ src/hooks/project/useProjectActions.ts
@@ -1,0 +1,141 @@
+import { useState } from "react"
+import { useNavigate } from "react-router-dom"
+import { useToast } from "@/components/common/ToastContext"
+import { projectApi } from "@/services/api/project-api"
+import { useProjectStore } from "@/stores/project-store"
+import type { Project } from "@/types/project"
+
+// 프로젝트 폼 데이터 타입
+type ProjectFormData = {
+  name: string
+  description: string
+}
+
+interface UseProjectActionsProps {
+  project: Project
+  onToggle?: (id: number) => void
+  onLeave?: (id: number) => void
+}
+
+export function useProjectActions({ project, onToggle, onLeave }: UseProjectActionsProps) {
+  const navigate = useNavigate()
+  const { refreshProjects } = useProjectStore()
+  const { addToast } = useToast()
+
+  // 다이얼로그 상태들
+  const [isToggleDialogOpen, setIsToggleDialogOpen] = useState(false)
+  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false)
+  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false)
+  const [isLeaveDialogOpen, setIsLeaveDialogOpen] = useState(false)
+  const [editLoading, setEditLoading] = useState(false)
+  const [deleteLoading, setDeleteLoading] = useState(false)
+
+  // 핸들러 함수들
+  const handleProjectClick = () => {
+    navigate(`/projects/${project.id}/ide`)
+  }
+
+  const handleToggleClick = () => setIsToggleDialogOpen(true)
+  const handleToggleConfirm = () => {
+    onToggle?.(project.id)
+    setIsToggleDialogOpen(false)
+  }
+
+  const handleEditClick = (e: React.MouseEvent) => {
+    e.stopPropagation()
+    setIsEditDialogOpen(true)
+  }
+
+  const handleEditConfirm = async (data: ProjectFormData) => {
+    setEditLoading(true)
+    try {
+      await projectApi.updateProject(project.id, {
+        projectName: data.name,
+        description: data.description,
+      })
+      setIsEditDialogOpen(false)
+      await refreshProjects()
+      addToast({
+        type: "success",
+        title: "프로젝트가 성공적으로 수정되었습니다.",
+        duration: 3000,
+      })
+    } catch (error) {
+      console.error("프로젝트 수정 실패:", error)
+      addToast({
+        type: "error",
+        title: "프로젝트 수정에 실패했습니다.",
+        duration: 3000,
+      })
+    } finally {
+      setEditLoading(false)
+    }
+  }
+
+  const handleDeleteClick = (e: React.MouseEvent) => {
+    e.stopPropagation()
+    setIsDeleteDialogOpen(true)
+  }
+
+  const handleDeleteConfirm = async () => {
+    setDeleteLoading(true)
+    try {
+      await projectApi.deleteProject(project.id)
+      setIsDeleteDialogOpen(false)
+      await refreshProjects()
+      addToast({
+        type: "success",
+        title: "project deleted successfully.",
+        duration: 3000,
+      })
+    } catch (_error) {
+      addToast({
+        type: "error",
+        title: "project deletion failed.",
+        duration: 3000,
+      })
+    } finally {
+      setDeleteLoading(false)
+    }
+  }
+
+  const handleLeaveClick = (e: React.MouseEvent) => {
+    e.stopPropagation()
+    setIsLeaveDialogOpen(true)
+  }
+
+  const handleLeaveConfirm = () => {
+    onLeave?.(project.id)
+    setIsLeaveDialogOpen(false)
+  }
+
+  return {
+    // 상태들
+    dialogStates: {
+      isToggleDialogOpen,
+      setIsToggleDialogOpen,
+      isEditDialogOpen,
+      setIsEditDialogOpen,
+      isDeleteDialogOpen,
+      setIsDeleteDialogOpen,
+      isLeaveDialogOpen,
+      setIsLeaveDialogOpen,
+    },
+    loadingStates: {
+      editLoading,
+      deleteLoading,
+    },
+    // 핸들러들
+    handlers: {
+      handleProjectClick,
+      handleToggleClick,
+      handleToggleConfirm,
+      handleEditClick,
+      handleEditConfirm,
+      handleDeleteClick,
+      handleDeleteConfirm,
+      handleLeaveClick,
+      handleLeaveConfirm,
+    },
+  }
+}
